langchain/requirements.txt
openai
fastapi
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
uvicorn[standard]
python-dotenv
pydantic>=2.12.0
duckduckgo-search>=6.0.0
ddgs>=0.1.0
requests>=2.31.0
pymongo>=4.6.0
motor>=3.3.0